#5383: Separate Falcon7b demo for grayskull, wormhole, t3000
